config: add ignore repos configuration
Add a new configuration that tracks repos that should be excluded. These
are stored as strings and support the same path resolution rules as our
repo roots (env, ~/, etc). This is useful for example if you have a huge
project like llvm that you don't care about tracking but also don't want
to move it out of your repo.